requirements.txt
Pillow
rply